requirements.txt
numpy
opencv-python
Pillow
PyQt5